Start splitting out QrCreator
Also adjusts EsLint for ES6 modules.
