+++ .eslintrc
@@ -2,7 +2,7 @@
     "root": true,
     "parserOptions": {
         "ecmaVersion": 2017,
-        "sourceType": "script",
+        "sourceType": "module",
         "ecmaFeatures": {
             "impliedStrict": true
         }
