+++ src/common/modules/GlobalConsts.js
@@ -1,0 +1,17 @@
+/**
+ * Specifies the long name of this add-on.,
+ *
+ * @readonly
+ * @type {string}
+ * @default
+ */
+export const ADDON_NAME = "Offline QR code generator"; // eslint-disable-line no-unused-vars
+
+/**
+ * Specifies the short name of this add-on.,
+ *
+ * @readonly
+ * @type {string}
+ * @default
+ */
+export const ADDON_NAME_SHORT = "Offline QR code";
