+++ src/common/modules/internal/MessageLevel.js
@@ -1,0 +1,14 @@
+/**
+ * Specifies the message level to use,
+ *
+ * @readonly
+ * @enum {int}
+ * @default
+ */
+export const MESSAGE_LEVEL = Object.freeze({
+    "ERROR": 3,
+    "WARN": 2,
+    "INFO": 1,
+    "LOADING": -2,
+    "SUCCESS": -3
+});
