+++ src/popup/qrcode.html
@@ -11,7 +11,7 @@
 
 		<script defer src="../common/lib/lodash/lodash.js" charset="utf-8"></script>
 		<script defer src="../common/common.js" charset="utf-8"></script>
-		<script defer src="qrcode.js" charset="utf-8"></script>
+		<script defer src="qrcode.js" type="module" charset="utf-8"></script>
 	</head>
 	<body class="center-container">
 		<div id="topMessageContainer" class="message-container center-container">
