+++ src/popup/qrcode.js
@@ -8,421 +8,10 @@
 // lodash
 /* globals throttle, isObject */
 
+import * as QrCreator from "./modules/QrCreator.js";
+
 /* globals */
 let initCompleted = false;
-
-const QrLibQrGen = (function () {
-    const me = {};
-
-    /* globals qrcodegen */
-    const QRC = qrcodegen.QrCode;
-
-    const qrBorder = 0;
-    let qrText;
-    let qrColor;
-    let qrErrorCorrection;
-
-    /**
-     * The type of QR code this library generates.
-     *
-     * @name QrLibQrGen.GENERATION_TYPE
-     * @type {string}
-     * @const
-     * @default
-     */
-    me.GENERATION_TYPE = "svg";
-
-    /**
-     * Generates an SVG element out of an SVG string.
-     *
-     * @name   QrLibQrGen.getSvgElement
-     * @function
-     * @private
-     * @param {string} svgString the SVG+XML string
-     * @returns {SVGSVGElement}
-     */
-    function getSvgElement(svgString) {
-        const svg = (new DOMParser()).parseFromString(svgString, "image/svg+xml"); // XMLDocument
-        const elSvg = svg.documentElement; // SVGSVGElement
-
-        // modify SVG
-        // transparent background
-        elSvg.querySelector("rect").setAttribute("fill", "transparent");
-        elSvg.querySelector("path").setAttribute("fill", qrColor);
-
-        return elSvg;
-    }
-
-    /**
-     * Set an option for the QR code.
-     *
-     * @name   QrLibQrGen.set
-     * @function
-     * @param {string} tag the common one you know from the outside, e.g. size
-     * @param {Object} value the value to set for this tag
-     * @returns {void}
-     */
-    me.set = function(tag, value) {
-        switch (tag) {
-        case "size":
-            // ignore as this property is not availavble in this lib
-            break;
-        case "text":
-            qrText = value;
-            break;
-        case "qrColor":
-            qrColor = value;
-            break;
-        case "qrBackgroundColor":
-            // ignore as this property is not availavble in this lib
-            break;
-        case "qrErrorCorrection":
-            switch (value) {
-            case "H":
-                qrErrorCorrection = QRC.Ecc.HIGH;
-                break;
-            case "Q":
-                qrErrorCorrection = QRC.Ecc.QUARTILE;
-                break;
-            case "M":
-                qrErrorCorrection = QRC.Ecc.MEDIUM;
-                break;
-            case "L":
-                qrErrorCorrection = QRC.Ecc.LOW;
-                break;
-            default:
-                throw new Error(`unknown error correction option passed: ${value}`);
-            }
-            break;
-        default:
-            throw new Error(`unknown tag passed to set: ${tag}`);
-        }
-    };
-
-    /**
-     * Return new QR code.
-     *
-     * @name   QrLibQrGen.getQr
-     * @function
-     * @returns {SVGSVGElement}
-     */
-    me.getQr = function() {
-        Logger.logInfo("generated new QrGen qr code");
-
-        const qrElem = QRC.encodeText(qrText, qrErrorCorrection);
-        const svgString = qrElem.toSvgString(qrBorder);
-
-        return getSvgElement(svgString);
-    };
-
-    return me;
-})();
-
-
-const QrLibKjua = (function () {
-    const me = {};
-
-    /* globals kjua */
-
-    /**
-     * The type of QR code this library generates.
-     *
-     * @name QrLibKjua.GENERATION_TYPE
-     * @type {string}
-     * @const
-     * @default
-     */
-    me.GENERATION_TYPE = "canvas";
-
-    /**
-     * The saved options for Kjua.
-     *
-     * @name QrLibKjua.kjuaOptions
-     * @private
-     */
-    let kjuaOptions;
-
-    /**
-     * How options need to be mapped from general options to kjua.
-     *
-     * format: generalOpt => kjua
-     *
-     * @name QrLibKjua.OPTIONS_MAP
-     * @private
-     */
-    const OPTIONS_MAP = Object.freeze({
-        "qrColor": "fill",
-        "qrBackgroundColor": "back",
-        "qrErrorCorrection": "ecLevel"
-    });
-
-    /**
-     * Resets all options for the QR code.
-     *
-     * @name   QrLibKjua.reset
-     * @function
-     * @returns {HTMLElement}
-     */
-    me.reset = function() {
-        kjuaOptions = {
-            // render method: 'canvas' or 'image'
-            render: "canvas",
-
-            // render pixel-perfect lines
-            crisp: true,
-
-            // minimum version: 1..40
-            minVersion: 1,
-
-            // error correction level: 'L', 'M', 'Q' or 'H'
-            ecLevel: "H",
-
-            // size in pixel
-            size: 200,
-
-            // pixel-ratio, null for devicePixelRatio
-            ratio: null,
-
-            // code color
-            fill: "#0c0c0d",
-
-            // background color
-            back: "#ffffff",
-
-            // roundend corners in pc: 0..100
-            rounded: 0,
-
-            // quiet zone in modules
-            quiet: 0,
-
-            // modes: 'plain', 'label' or 'image'
-            mode: "plain",
-
-            // label/image size and pos in pc: 0..100
-            mSize: 30,
-            mPosX: 50,
-            mPosY: 50
-        };
-    };
-
-    /**
-     * Set an option for the QR code.
-     *
-     * @name   QrLibKjua.set
-     * @function
-     * @param {string} tag the common one you know from the outside, e.g. size
-     * @param {Object} value the value to set for this tag
-     * @returns {void}
-     */
-    me.set = function(tag, value) {
-        if (OPTIONS_MAP.hasOwnProperty(tag)) {
-            tag = OPTIONS_MAP[tag];
-        }
-
-        // TODO: should it reject invalid values?
-
-        kjuaOptions[tag] = value;
-    };
-
-    /**
-     * Return new QR code.
-     *
-     * @name   QrLibKjua.getQr
-     * @function
-     * @returns {HTMLElement}
-     */
-    me.getQr = function() {
-        Logger.logInfo("generated new qr kjua code", kjuaOptions);
-        return kjua(kjuaOptions);
-    };
-
-    /**
-     * Init connector module.
-     *
-     * @name   QrLibKjua.init
-     * @function
-     * @returns {void}
-     */
-    me.init = function() {
-        me.reset();
-    };
-
-    return Object.freeze(me);
-})();
-
-// abstracts away all specific handling of QR code library
-const QrCreator = (function () {
-    const me = {};
-
-    let initFinished = false;
-    let qrCodeLib = null;
-
-    // by default everything has "been changed" (i.e. nothing has been generated yet)
-    const changedValues = new Set("text", "color", "size");
-
-    /**
-     * Provide connection to library and get QR code with current options.
-     *
-     * @name   QrCreator.getQrCodeFromLib
-     * @function
-     * @private
-     * @returns {HTMLElement}
-     */
-    function getQrCodeFromLib() {
-        return qrCodeLib.getQr();
-    }
-
-    /**
-     * Displays the QR code when options have been set.
-     *
-     * @name   QrCreator.generate
-     * @function
-     * @returns {void}
-     */
-    me.generate = function() {
-        if (!initFinished) {
-            Logger.logWarning("QrCreator.generate called, but init not yet finished. Abort.");
-            return;
-        }
-
-        // special shortcuts for SVG output when text does not need to be regenerated
-        if (qrCodeLib.GENERATION_TYPE === "svg" && !changedValues.has("text")) {
-            // color won't be changed
-            // size does not need adjustment for SVGs
-
-            return;
-        }
-
-        UserInterface.replaceQr(getQrCodeFromLib());
-
-        changedValues.clear();
-    };
-
-    /**
-     * Sets the size of the QR code.
-     *
-     * @name   QrCreator.setSize
-     * @function
-     * @param  {int} size
-     * @returns {void}
-     */
-    me.setSize = function(size) {
-        if (size <= 1) {
-            Logger.logError("tried to create QR code with invalid size of 0 or smaller");
-            return;
-        }
-
-        qrCodeLib.set("size", size);
-    };
-
-    /**
-     * Sets the text for the QR code.
-     *
-     * Note that this also triggers all user interface actions to display the
-     * test in a nice way. (e.g. selection and scrolling)..
-     * As such, it is not a good idea for live updating the text. To only set
-     * the option for the QR code itself, use {@link setTextInternal()}.
-     *
-     * @name   QrCreator.setText
-     * @function
-     * @param  {string} text
-     * @returns {void}
-     */
-    me.setText = function(text) {
-        me.setTextInternal(text);
-        UserInterface.setQrInputFieldText(text);
-    };
-
-    /**
-     * Sets the text for the QR code, but does not modify the UI.
-     *
-     * Usually this should not be used, as it can cause an inconsistent display.
-     *
-     * @name   QrCreator.setTextInternal
-     * @function
-     * @param  {string} text
-     * @returns {void}
-     */
-    me.setTextInternal = function(text) {
-        changedValues.add("text");
-        qrCodeLib.set("text", text);
-    };
-
-    /**
-     * Generates a QR code from a given tab.
-     *
-     * @name   QrCreator.generateFromTab
-     * @function
-     * @param  {browser.tabs} tab
-     * @returns {void}
-     */
-    me.generateFromTab = function(tab) {
-        if (tab.url === undefined) {
-            throw new Error("URL not yet available.");
-        }
-
-        me.setText(tab.url);
-        me.generate();
-    };
-
-    /**
-     * Generates a QR code for multiple tabs.
-     *
-     * Attention: Currently just uses the first tab, only!
-     *
-     * @name   QrCreator.generateFromTabs
-     * @function
-     * @param  {browser.tabs} tabs tabs passed from browser.tabs
-     * @returns {void}
-     */
-    me.generateFromTabs = function(tabs) {
-        me.generateFromTab(tabs[0]);
-    };
-
-    /**
-     * Returns the type of the generated QR code.
-     *
-     * @name   QrCreator.getGenerationType
-     * @function
-     * @returns {Promise}
-     */
-    me.getGenerationType = async function() {
-        await qrCreatorInit; // module needs to be initiated
-
-        return qrCodeLib.GENERATION_TYPE;
-    };
-
-    /**
-     * Initiates module.
-     *
-     * @name   QrCreator.init
-     * @function
-     * @returns {Promise}
-     */
-    me.init = function() {
-        // get all settings
-        return AddonSettings.get().then((settings) => {
-            switch (settings.qrCodeType) {
-            case "svg":
-                qrCodeLib = QrLibQrGen;
-                break;
-            case "canvas":
-                qrCodeLib = QrLibKjua;
-                break;
-            default:
-                throw new Error("invalid QR code type setting");
-            }
-
-            qrCodeLib.set("qrColor", settings.qrColor);
-            qrCodeLib.set("qrBackgroundColor", settings.qrBackgroundColor);
-            qrCodeLib.set("qrErrorCorrection", settings.qrErrorCorrection);
-
-            initFinished = true;
-        });
-    };
-
-    return me;
-})();
 
 const UserInterface = (function () {
     const me = {};
@@ -1096,7 +685,6 @@ const BrowserCommunication = (function () {
 // init modules
 const queryBrowserTabs = browser.tabs.query({active: true, currentWindow: true});
 AddonSettings.loadOptions();
-QrLibKjua.init();
 BrowserCommunication.init();
 const qrCreatorInit = QrCreator.init();
 const userInterfaceInit = UserInterface.init();
